Use ERC20Upgradeable to initialize token name and symbol (#9)
The previous implementation of `SUPTB` was setting the token `name` and `symbol` in the storage of the implementation contract, so they couldn't be accessed in the proxy.

This PR updates the contract to extend `ERC20Upgradeable` instead of `ERC20` in order to initialize `name` and `symbol` in the proxy using the `initialize` function. 

An alternative approach would be to set the `name` and `description` as immutables in the constructor, but these values would need to be defined as `byte32`'s instead of strings. This is quite limiting, given the fact that the token name could potentially be much longer than 32 characters.